# lib/api-zod/src/generated/types/walletConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNILIV Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WalletConfigBody {
  minimumBalance?: number | null;
  lowBalanceAlert?: number | null;
  isEnabled?: boolean | null;
  topupNotes?: string | null;
}