# lib/api-zod/src/generated/types/walletConfigDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNILIV Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WalletConfigDto {
  id: string;
  propertyId: string;
  minimumBalance: number;
  lowBalanceAlert: number;
  isEnabled: boolean;
  topupNotes?: string | null;
  createdAt?: Date;
  updatedAt?: Date;
}