# lib/api-zod/src/generated/types/walletDetailDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNILIV Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WalletDetailDto {
  id: string;
  residentId: string;
  balance: number;
  isActive: boolean;
  residentName: string;
  walletEnabled: boolean;
  transactionCount: number;
  createdAt?: Date;
  updatedAt?: Date;
}