# lib/api-zod/src/generated/types/walletPayBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNILIV Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WalletPayBody {
  amount: number;
  description: string;
  notes?: string | null;
  referenceId?: string | null;
  referenceType?: string | null;
}